jaja

+++ src/main/java/ap/hackathon/augmentedprivacy/domain/Account.java
@@ -2,13 +2,13 @@ package ap.hackathon.augmentedprivacy.domain;
 
 import lombok.Data;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class Account {
 
-    private BigInteger balance;
+    private BigDecimal balance;
     private List<Transaction> transactions;
 
 }
